+++ src/lib.rs
@@ -10,15 +10,7 @@ mod tests {
     use serde_json::json;
 
     use crate::ilert::ILert;
-    use crate::ilert_builders::{
-        UserApiResource,
-        EventApiResource,
-        ScheduleApiResource, 
-        HeartbeatApiResource,
-        ILertEventType,
-        ILertPriority,
-        EventImage
-    };
+    use crate::ilert_builders::{UserApiResource, EventApiResource, ScheduleApiResource, HeartbeatApiResource, ILertEventType, ILertPriority, EventImage, EventComment};
 
     #[test]
     fn init() -> () {
@@ -56,7 +48,7 @@ mod tests {
     }
 
     #[test]
-    fn create_and_resolve_event_test() {
+    fn create_comment_and_resolve_event_test() {
 
         let mut client = ILert::new_with_opts(Some("http://localhost:8080"), Some(10)).unwrap();
 
@@ -64,28 +56,43 @@ mod tests {
             .post()
             .event_with_details(
                 "il1api0220953b09684c9e4fe8972f0d5d8c9cde78d79b6cc8fd",
-                ILertEventType::ALERT, "Host srv/mail01 is CRITICAL",
+                ILertEventType::ALERT,
+                Some("Host srv/mail01 is CRITICAL".to_string()),
                 Some("bratwurst".to_string()),
                 Some("some detail message".to_string()),
                 Some(ILertPriority::LOW),
                 Some(vec![EventImage::new("https://i.giphy.com/media/VRhsYYBw8AE36/giphy.webp")]),
                 Some(vec![]),
-                Some(json!({"hehe": "test"}))
+                Some(json!({"hehe": "test"})),
+                None
             )
             .execute()
             .unwrap();
 
-        assert_eq!(event_result.status, 200);
+        assert_eq!(event_result.status, 202);
+
+        let event_comment_result = client
+            .post()
+            .event_with_comment(
+                "il1api0220953b09684c9e4fe8972f0d5d8c9cde78d79b6cc8fd",
+                Some("bratwurst".to_string()),
+                Some(vec![EventComment::new("Peter Parker",
+                                            "a comment ![alt text picture](https://i.giphy.com/media/VRhsYYBw8AE36/giphy.webp) salut")])
+            )
+            .execute()
+            .unwrap();
+
+        assert_eq!(event_comment_result.status, 202);
 
         let resolve_result = client
             .post()
             .event("il1api0220953b09684c9e4fe8972f0d5d8c9cde78d79b6cc8fd",
-                   ILertEventType::RESOLVE, "Host srv/mail01 is CRITICAL",
+                   ILertEventType::RESOLVE, None,
                     Some("bratwurst".to_string()))
             .execute()
             .unwrap();
 
-        assert_eq!(resolve_result.status, 200);
+        assert_eq!(resolve_result.status, 202);
     }
 
     #[test]